Move ssr.routes to routes folder
